benchmarks/requirements.txt
numpy
pytest-benchmark